refact - removing TIMEOUT const
