--- fastrpa/settings.py
@@ -1,1 +1,0 @@
-TIMEOUT = 15
